fix: remove redundant (int) cast on gfrv() return in host_reindex
gfrv() validates with FILTER_VALIDATE_INT by default; the cast is
redundant and inconsistent with call sites elsewhere in host.php.

+++ host.php
@@ -199,7 +199,7 @@ switch (grv('action')) {
 function host_reindex() : void {
 	$start = microtime(true);
 
-	$host_id = (int) gfrv('host_id');
+	$host_id = gfrv('host_id');
 	shell_exec(cacti_escapeshellcmd(read_config_option('path_php_binary')) . ' -q ' . cacti_escapeshellarg(CACTI_PATH_CLI . '/poller_reindex_hosts.php') . ' --qid=all --id=' . $host_id);
 
 	$end = microtime(true);
